restructure dataService files
