+++ src/DataService/localDataService.js
@@ -1,0 +1,5 @@
+function saveData(data) {
+  localStorage.setItem('formData', JSON.stringify(data));
+}
+
+export { saveData };
